*fixed bug with XP *applied patch to Riding Skill by andy012345

+++ src/game/Stats.h
@@ -169,7 +169,7 @@ inline uint32 CalculateXpToGive(Unit *pVictim, Unit *pAttacker)
 		if(victimI->Type == CRITTER)
 			return 0;
 	//not wowwikilike but more balanced
-	if (pVictim->getLevel()-pAttacker->getLevel()>10)
+	if ((int32)pVictim->getLevel()-(int32)pAttacker->getLevel()>10)
 		return 0;
 
 	uint32 max_level = sWorld.Expansion1LevelCap;
